WordloomBackend/api/requirements.txt

fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.0
sqlalchemy>=2.0
python-multipart>=0.0.9
passlib[bcrypt]>=1.7
PyJWT==2.8.0
psycopg2-binary>=2.9
